refactor : techset 도멘인 데이터 레이어 모듈 리팩토링
