+++ lib/features/tech_set/data_source/local/tech_set_local_data_source.dart
@@ -1,6 +1,7 @@
-import 'package:techtalk/features/tech_set/data_source/remote/models/job_model.dart';
+import 'package:techtalk/features/tech_set/repositories/enums/job.enum.dart';
 
 abstract interface class TechSetLocalDataSource {
+  /// 개발 직군 리스트 호출
   Future<List<Job>> getJobs();
 
   /// 스킬(Json String) 호출
