--- lib/features/tech_set/data_source/remote/tech_set_data_source.dart
@@ -1,5 +1,0 @@
-import 'package:techtalk/features/tech_set/data_source/remote/models/job_model.dart';
-
-abstract interface class TechSetDataSource {
-  Future<List<Job>> getJobs();
-}
