--- lib/features/tech_set/data_source/remote/tech_set_data_source_impl.dart
@@ -1,15 +1,0 @@
-// import 'dart:async';
-//
-//
-// import 'package:techtalk/features/tech_set/tech_set.dart';
-//
-// final class TechSetRemoteDataSourceImpl implements TechSetDataSource {
-//   @override
-//   Future<List<JobEntity>> getJobs() async {
-//     final cachedJobs = await FirestoreJobsRef.collection().get();
-//
-//     return [
-//       ...cachedJobs.docs.map((e) => e.data().toEntity()),
-//     ];
-//   }
-// }
