+++ lib/features/tech_set/repositories/enums/job.enum.dart
