+++ lib/features/tech_set/usecases/get_jobs_use_case.dart
@@ -1,4 +1,4 @@
-import 'package:techtalk/features/tech_set/data_source/remote/models/job_model.dart';
+import 'package:techtalk/features/tech_set/repositories/enums/job.enum.dart';
 import 'package:techtalk/features/tech_set/tech_set.dart';
 
 final class GetJobsUseCase {
