+++ lib/presentation/pages/sign_up/providers/sign_up_jobs_provider.dart
@@ -1,5 +1,5 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
-import 'package:techtalk/features/tech_set/data_source/remote/models/job_model.dart';
+import 'package:techtalk/features/tech_set/repositories/enums/job.enum.dart';
 
 part 'sign_up_jobs_provider.g.dart';
